Initial structure for dagster Java module
