+++ java-dagster-pipes/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Oct 01 14:12:37 MSK 2024
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.8-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
