+++ java-dagster-pipes/src/main/java/dagster/pipes/CauseContext.java
@@ -1,11 +1,13 @@
-package io.quicktype;
+package dagster.pipes;
 
 import com.fasterxml.jackson.annotation.*;
 
 /**
+ * exception that being handled when this exception was raised
+ *
  * exception that explicitly led to this exception
  */
-public class ExceptionCause {
+public class CauseContext {
     private CauseCause cause;
     private CauseContext context;
     private String message;
