+++ java-dagster-pipes/src/main/java/dagster/pipes/Converter.java
@@ -12,15 +12,14 @@
 //
 //     Metadata data = Converter.fromJsonString(jsonString);
 
-package io.quicktype;
+package dagster.pipes;
 
 import java.io.IOException;
 import com.fasterxml.jackson.databind.*;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import java.util.*;
-import java.time.LocalDate;
+
 import java.time.OffsetDateTime;
 import java.time.OffsetTime;
 import java.time.ZoneOffset;
