+++ java-dagster-pipes/src/main/java/dagster/pipes/ExceptionContext.java
@@ -1,4 +1,4 @@
-package io.quicktype;
+package dagster.pipes;
 
 import com.fasterxml.jackson.annotation.*;
 
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.*;
  *
  * exception that explicitly led to this exception
  */
-public class CauseContext {
+public class ExceptionContext {
     private CauseCause cause;
     private CauseContext context;
     private String message;
