+++ java-dagster-pipes/src/main/java/dagster/pipes/Metadata.java
@@ -1,7 +1,6 @@
-package io.quicktype;
+package dagster.pipes;
 
 import com.fasterxml.jackson.annotation.*;
-import java.util.Map;
 
 public class Metadata {
     private RawValue rawValue;
