+++ java-dagster-pipes/src/main/java/dagster/pipes/RawValue.java
@@ -1,7 +1,7 @@
-package io.quicktype;
+package dagster.pipes;
 
 import java.io.IOException;
-import java.io.IOException;
+
 import com.fasterxml.jackson.core.*;
 import com.fasterxml.jackson.databind.*;
 import com.fasterxml.jackson.databind.annotation.*;
