+++ java-dagster-pipes/src/main/java/dagster/pipes/Type.java
@@ -1,4 +1,4 @@
-package io.quicktype;
+package dagster.pipes;
 
 import java.io.IOException;
 import com.fasterxml.jackson.annotation.*;
